Capture and restore the process working directory (#4)
It is possible for colcon extensions of any kind to change the process
working directory, which could affect the sub-verbs invoked as part of
an alias. This change captures the working directory as early as
possible and restores it prior to each sub-verb invocation, which is
closer to what the behavior would be had the sub-verbs been invoked
manually from the shell.