Use only the base name of the program in argparse help
Showing the "full" program name where appropriate in some parts of the
help output is useful, but in this context, we're really referring more
to the program's title in a more abstract sense. Since we don't want to
re-invoke colcon_core.command.get_prog_name, we can narrow our current
use of the argparse reported prog to only the basename which will
usually give us exactly what we want.